Vivek_Trivedi/Neural RAG Research Intelligence/requirements.txt
gradio==5.49.1
python-dotenv==1.2.1
requests==2.32.5
lancedb==0.25.2
langchain==1.0.5
langchain-core==1.0.4
langchain-community==0.4.1
langchain-openai==1.0.2
langgraph==1.0.2
langgraph-sdk==0.2.9
llama-index-core==0.14.7
llama-index-vector-stores-lancedb==0.4.1
llama-index-embeddings-huggingface==0.6.1
llama-index-llms-openrouter==0.4.2
sentence-transformers==2.7.0
tavily-python==0.7.12
